TMP COMMIT

+++ backend/src/deeds/deeds.service.ts
@@ -1,9 +1,10 @@
-import { Injectable } from "@nestjs/common";
+import { Injectable, NotFoundException } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Repository } from "typeorm";
 import { Deed } from "./deed.entity";
 import { User } from "../users/user.entity";
 import { UsersService } from "../users/users.service";
+import { DeedDto } from "./dto";
 
 @Injectable()
 export class DeedsService {
@@ -13,14 +14,18 @@ export class DeedsService {
     private readonly usersService: UsersService,
   ) {}
 
-  async createDeed(userId: number, title: string, description: string) {
+  async createDeed(
+    userId: number,
+    title: string,
+    description: string,
+  ): Promise<DeedDto> {
     const user = await this.usersService.findById(userId);
     const deed = this.deedsRepository.create({ title, description, user });
     await this.deedsRepository.save(deed);
     return { title: deed.title, description: deed.description, id: deed.id };
   }
 
-  async findAllByUser(userId: number): Promise<Deed[]> {
+  async findAllByUser(userId: number): Promise<DeedDto[]> {
     const user = await this.usersService.findById(userId);
     return this.deedsRepository.find({
       where: { user },
@@ -36,14 +41,21 @@ export class DeedsService {
     deedId: number,
     title: string,
     description: string,
-  ): Promise<Deed> {
+  ): Promise<DeedDto> {
     const deed = await this.deedsRepository.findOne({ where: { id: deedId } });
+    if (!deed) throw new NotFoundException(`Deed with id ${deedId} not found`);
     deed.title = title;
     deed.description = description;
     return this.deedsRepository.save(deed);
   }
 
   async deleteDeed(deedId: number): Promise<void> {
+    const deed = await this.deedsRepository.findOne({ where: { id: deedId } });
+
+    if (!deed) {
+      throw new NotFoundException(`Deed with ID ${deedId} not found`);
+    }
+
     await this.deedsRepository.delete(deedId);
   }
 }
